Rpc check routine for valid lie basis (#232)
* Add a function to test the validity of Lie basis data

* Add functions to compute the size of the Hall set for use in constructing hall sets

* Add flaxibility to choose Hall set definition of Hall set used for checking routines.

* cast to int to stop compiler warning

* remove unused variable

+++ roughpy_compute/common/bitmask.hpp
@@ -21,7 +21,7 @@ public:
 
     constexpr UInt operator[](int idx) const noexcept
     {
-        assert(idx < 8 * sizeof(UInt));
+        assert(idx < static_cast<int>(8 * sizeof(UInt)));
         return (base_ >> idx) & UInt{1};
     }
